+++ block/block.go
@@ -16,7 +16,7 @@ import (
 // - block height is the expected block height on the store (height + 1).
 // - block height is the expected block height on the app (last block height + 1).
 func (m *Manager) applyBlock(ctx context.Context, block *types.Block, commit *types.Commit, blockMetaData blockMetaData) error {
-	// TODO (#330): allow genesis block with heigh > 0 to be applied.
+	// TODO (#330): allow genesis block with height > 0 to be applied.
 	// TODO: add switch case to have defined behavior for each case.
 	//validate block height
 	if block.Header.Height != m.store.NextHeight() {
